magneton/requirements.txt
asgiref==3.7.2
avro==1.11.2
boto3==1.28.5
botocore==1.31.5
Django==4.2.3
django-cors-headers==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
numpy==1.25.1
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
